fix: Prolific - public link deleted, user is unable to submit
